FE Refactoring, working validation on closed task form
