+++ myexaminer-frontend/src/components/exam/TaskForm.js
@@ -1,25 +1,14 @@
 import React, { useState } from 'react'
-import { useForm, Controller, useFieldArray, useFormContext } from "react-hook-form";
-import { Box, Button, FormControlLabel, TextField } from "@material-ui/core";
+import { useForm, Controller, useFieldArray } from "react-hook-form";
+import { Box, Button, TextField } from "@material-ui/core";
 import styles from "components/group/group.module.css";
 import { createLessonUrl, lessonIdUrl } from "router/urls";
 import authHeader from "services/auth-header";
 import CheckIcon from '@material-ui/icons/Check';
 import CloseIcon from '@material-ui/icons/Close';
-
 import { makeStyles } from '@material-ui/core/styles';
 import Radio from '@material-ui/core/Radio';
-import RadioGroup from '@material-ui/core/RadioGroup';
-import FormHelperText from '@material-ui/core/FormHelperText';
-
-const isNumeric = (num) => {
-  return !isNaN(num);
-}
-
-const isWholeNumber = (num) => {
-  num = parseInt(num);
-  return Number.isInteger(num);
-}
+import { isWholeNumber, isNumeric } from "utils/validationUtils";
 
 const useStyles = makeStyles((theme) => ({
   formControl: {
@@ -41,7 +30,7 @@ export default function TaskForm(props) {
       points: 1,
       // ??? fill: "hello <blank>!", answers:["1,F", "2,T", ...]
       answers: [
-        { text: "", value: "F" },
+        { text: "", value: "T" },
         { text: "", value: "F" },
         { text: "", value: "F" },
         { text: "", value: "F" }
@@ -116,15 +105,14 @@ export default function TaskForm(props) {
       submitText = "Zatwierdź";
   }
 
-// ==================================================
-  const [value2, setValue2] = useState('');
   const [radioValue, setRadioValue] = useState(0);
-  const [error, setError] = useState(false);
 
-  const handleRadioChange = (event) => {
-    setValue2(event.target.value);
-    setError(false);
-  };
+  const { fields } = useFieldArray(
+    {
+      control,
+      name: "answers"
+    }
+  );
 
   const handleRadio = (selected) => {
     getValues()["answers"].forEach((_value, index) => {
@@ -134,16 +122,6 @@ export default function TaskForm(props) {
       setValue(field, newValue);
     })
   }
-
-// ==================================================
-  const { fields, append, prepend, remove, swap, move, insert } = useFieldArray(
-    {
-      control,
-      name: "answers"
-    }
-  );
-
-  // ==================================================
 
   return (
     <form onSubmit={handleSubmit(action)} className={styles.lessonForm}>
@@ -197,16 +175,10 @@ export default function TaskForm(props) {
             }
           }}
         />
-        {/* ======= */}
-
         
-        {/* ======= */}
-        {/* <RadioGroup aria-label="answers" name="answers" value={radioValue} > */}
-          {/* onChange={(e) => { setRadioValue(e.target.value); handleRadio(e.target.value); }}> */}
         {fields.map((item, index) => {
           return (
             <Box key={item.id} display="flex" style={{paddingBottom: "20px", width: "50%"}}>
-              
               <Controller
                 control={control}
                 name={`answers.${index}.value`}
@@ -214,44 +186,38 @@ export default function TaskForm(props) {
                   <Radio
                     value={index}
                     name={`answers.${index}.value`}
-                    // name="answers"
-                    // defaultChecked={index === radioValue}
                     checked={index == radioValue}
                     onChange={(event) => {
-                      // onChange(event.target.value);
                       setRadioValue(event.target.value);
-                      console.log(event.target.value)
                       handleRadio(index);
                     }}
                     {...rest}
                   />
-              }
+                }
               />
               
               <Controller
-                  control={control}
-                  name={`answers.${index}.text`}
-                  defaultValue={item.text}
-                  render={({ field }) =>
-                    <TextField
-                      variant="outlined"
-                      label={`Odpowiedź nr ${index+1}`}
-                      // error={errors.answers.index.text}
-                      fullWidth
-                      // helperText={ errors.answers.index.text ? errors.answer.index.text?.message : null }
-                      {...field}
-                    />
-                  }
-                  rules={{
-                    required: "Wypełnij to pole"
-                  }}
-                // />
-              // } />
+                control={control}
+                name={`answers.${index}.text`}
+                defaultValue={item.text}
+                render={({ field }) =>
+                  <TextField
+                    variant="outlined"
+                    label={`Odpowiedź nr ${index + 1}`}
+                    fullWidth
+                    error={errors?.answers?.[index]?.text}
+                    helperText={ errors?.answers?.[index]?.text ? errors?.answers[index]?.text?.message : null }
+                    
+                    {...field}
+                  />
+                }
+                rules={{
+                  required: "Wypełnij to pole"
+                }}
               />
-              </Box>
+            </Box>
           );
         })}
-        {/* </RadioGroup> */}
         
       </Box>
       <Box display="flex" justifyContent="flex-end">
