+++ myexaminer-frontend/src/utils/validationUtils.js
@@ -1,0 +1,10 @@
+const isNumeric = (num) => {
+  return !isNaN(num);
+};
+
+const isWholeNumber = (num) => {
+  num = parseInt(num);
+  return Number.isInteger(num);
+};
+
+export { isNumeric, isWholeNumber };
